backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
tortoise-orm>=0.20.0
aiosqlite>=0.19.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aerich>=0.7.2
